Generalise ChannelWrapper in tools package, including base impl
